Add check for prerequisites programs
